+++ NEWS.md
@@ -1,3 +1,11 @@
+# simplevis 5.0.42
+
+* Added alpha arguments to sf and stars.
+
+# simplevis 5.0.41
+
+* Added alpha arguments to line.
+
 # simplevis 5.0.40
 
 * Added alpha arguments to point.